backend/requirements.txt
    # backend/requirements.txt

    fastapi
    uvicorn[standard]
    pydantic

    openai
    langchain
    langgraph
    langchain-openai
    langchain-core
    yfinance
    pandas
    ta
    numpy
    matplotlib
    gunicorn
    alpha_vantage
    python-dotenv
    httpx

    # Added for chart generation
    # langsmith # Optional
